pyproject.toml
[tool.poetry]
name = "download_arxiv"
version = "0.1.6"
description = ""
authors = ["Shun Kiyono <kiyono@ecei.tohoku.ac.jp>"]
readme = "README.md"
license = "MIT"

[tool.poetry.dependencies]
python = "^3.6"
click = "^7.0"
arxiv = "^0.3.1"
python-dateutil = "^2.8"
logzero = "^1.5"
pybtex = "^0.22.2"
lxml = "^4.5.0"

[tool.poetry.dev-dependencies]
ipdb = "^0.12.0"

[tool.poetry.scripts]
arxiv = "download_arxiv:main"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"


